make nice svg animation for prize2
