+++ source/index.html
@@ -194,19 +194,12 @@
           </div>
           <ul class="prizes__list">
             <li class="prizes__item prizes__item--journeys js-prizes-item-journeys">
-
               <p class="prizes__desc">
                 <b>3</b>
                 <span>загадочных путешествия</span>
               </p>
             </li>
-            <li class="prizes__item prizes__item--cases">
-              <div class="prizes__icon">
-                <picture>
-                  <source srcset="img/prize2-mob.svg" media="(orientation: portrait)">
-                  <img src="img/prize2.svg" alt="">
-                </picture>
-              </div>
+            <li class="prizes__item prizes__item--cases js-prizes-item-cases">
               <p class="prizes__desc">
                 <b>7</b>
                 <span>надежных чемоданов</span>
