+++ source/js/modules/menu.js
@@ -1,4 +1,4 @@
-import browser from "swiper/src/modules/browser/browser";
+import prizes from '../view/prizes';
 
 export default () => {
   let header = document.querySelector(`.js-header`);
@@ -31,33 +31,14 @@ export default () => {
       const isCurrentLocationStory = window.location.toString().includes(`#story`);
 
       const prizesItemYourneys = document.querySelector(`.js-prizes-item-journeys`);
+      const prizesItemCases = document.querySelector(`.js-prizes-item-cases`);
 
       if (isCurrentMenuLinkRules && !!rulesLink && rulesLink.classList.contains(`active`)) {
         rulesLink.classList.remove(`active`);
       }
 
       if (isCurrentMenuLinkPrizes) {
-        const journeysTemplate = document.createElement(`div`);
-        journeysTemplate.innerHTML = `<div class="prizes__icon prizes__icon--from">
-                <div class="prizes__icon-clouds">
-                  <img src="./img/prize1-clouds.svg" alt="">
-                </div>
-                <div class="prizes__icon-airship">
-                  <div class="prizes__icon-airship-dirigible">
-                    <img src="./img/prize1-airship.svg" alt="">
-                  </div>
-                  <div class="prizes__icon-airship-emissions">
-                    <img src="./img/prize1-airship-emissions.svg" alt="">
-                  </div>
-                </div>
-                <img src="./img/prize1-land.svg" alt="">
-              </div>
-              <div class="prizes__icon prizes__icon--to">
-                <img src="./img/prize1-ski.svg" alt="">
-                <img src="./img/prize1-chair.svg" alt="">
-                <img src="./img/prize1-branch.svg" alt="">
-              </div>`;
-        prizesItemYourneys.prepend(journeysTemplate);
+        prizes();
 
         if (isCurrentLocationStory) {
           evt.preventDefault();
@@ -72,6 +53,7 @@ export default () => {
         }
       } else {
         prizesItemYourneys.firstElementChild.remove();
+        prizesItemCases.firstElementChild.remove();
       }
     });
   }
